data_analyse-dev/requirements.txt
streamlit
streamlit-chat==0.1.1
streamlit-feedback==0.1.3
langchain==0.1.11
langchain-community==0.0.27
langchain-experimental==0.0.53
langchain-openai==0.0.8
langchain-pinecone==0.0.3
unstructured
#unstructured-inference==0.7.24
#unstructured.pytesseract==0.3.12
elasticsearch==8.12.1
python-dotenv
matplotlib
seaborn==0.13.2
numpy==1.23.5
scipy
libmagic==1.0
tesseract==0.1.3
#pinecone-client==3.1.0
#wikipedia==1.4.0
scikit-learn
#sentence_transformers
pillow_heif==0.15.0
altair==4.1.0
#chromadb==0.4.24
#pikepdf==8.13.0
#pypdf==4.1.0
